Make a proper setContent message and move old code to it's own method
The current #setContent method does not work as the name suggests, and does not actually set the content, rather append the supplied content to the beginning of the current content. I've made a new method that keeps this behavior, and changed the setContent method to what it should actually do.